ensure that this role is only used on ubuntu
This is just a small nicety to give good feedback if the role is
accidentally used on another distro.

It would fail anyhow on any other distro, but this way we do not change
anything on the target (which might not make sense in the same way on
another distribution) and the user gets a much better error message,
especially if they are not familiar with ubuntu and its commands.