translate instructions from ros wiki to ansible
I have translated the
[installation instructions](https://docs.ros.org/en/humble/Installation/Ubuntu-Install-Debians.html)
for humble on ubuntu jammy into the ansible playbook.

I have omitted the checks for the locale and the universe repo as they
are not strictly necessary and as I want to add them as a separate
commit.

I did not install `curl` and `lsb-release` as I seem to not
need these tools when using ansible. `gnupg` seems to be installed in
the podman image that I use for testing, therefore I do not know whether
it is needed. Although the output of `dpkg -l gnupg` is a bit weird:
```
$ podman run ubuntu-ansible-ready:jammy dpkg -l gnupg
Desired=Unknown/Install/Remove/Purge/Hold
| Status=Not/Inst/Conf-files/Unpacked/halF-conf/Half-inst/trig-aWait/Trig-pend
|/ Err?=(none)/Reinst-required (Status,Err: uppercase=bad)
||/ Name           Version      Architecture Description
+++-==============-============-============-=================================
un  gnupg          <none>       <none>       (no description available)
```

We need `ca-certificates` because otherwise we will get an error that
we cannot download the ros keyring, as we cannot verify the https
certificate.

I chose `rolling` as the default distro, because it will never get out
of date ;) (if I i.e. chose `humble` I would need to update this at
least when `humble` is end of life)

I chose `desktop` as the default metapackage, as this is the most
comprehensive package, which is I think a good default for people who
don't know what to choose.