test the role
Try to install all non end of life ros2 distributions on their
supported ubuntu versions. This way we can be relatively sure that this
role will work for "any" ros2 distribution, as the README claims.

I only install `ros-base` since this already takes quite some time
(`molecule test` takes about 10 minutes on my setup) and we don't want
to test whether ros2 installs flawlessly, but only if our playbook works
as expected and for that `ros-base` should be enough.

I have added the instructions on how to test to the README to make it
easy to replicate.

We need a custom image since ansible requires the target to have
`python` installed and since we use `become` we also need `sudo`.