Check that the user has set up a supported locale
The installation instructions strongly recommend you setting a locale
that supports `UTF-8`.

I do not want to select the locale for the user, as this is really up to
the users preference.
Therefore I just check if the locale is set.

Unfortunately I cannot use `localectl` as it depends on `systemd` and I
do not want to start it in the test containers (and on `docker` it would
be even more difficult, if one wants to test with `docker` instead).
Therefore I just check for `LANG`, as it is set by the normal desktop
install and the instructions on the ROS page would also make the user
set `LANG`.

Should anyone not set `LANG` but all of the other `LC_*` variables we
can still change it later on.

On the test container we now need to set the `LANG` variable.
Setting `LANG` in `/etc/default/locale` does not work, since
`meta: reset_connection` is not implemented for this connection and the
file is only sourced when creating the connection.
I have also tried adding an extra `gather_facts: yes`, but that does not
work either.
Therefore I just added the `environment: {LANG: en_US.UTF-8}` to every
playbook.

While generating the locales was not strictly necessary to pass the test
I still added it for completeness sake.

See also:
https://docs.ros.org/en/humble/Installation/Ubuntu-Install-Debians.html#set-locale