Check if `universe` is enabled
ros2 requires `universe` to be enabled (see also the
[installation instructions](https://docs.ros.org/en/humble/Installation/Ubuntu-Install-Debians.html#setup-sources)),
so we need to ensure that it is enabled.

Unfortunately ansibles `apt_repository`
[does not support enabling components](https://github.com/ansible/ansible/issues/48714)
directly.

Also since one could use a mirror it is not as simple as just checking
whether the
[lines from this comment](https://github.com/ansible/ansible/issues/48714#issuecomment-441911493)
are in `/etc/apt/sources.list`. Therefore I used the heuristic that was
also provided in the installation instructions.

That way we will not have any false positives and install maybe unwanted
mirrors that make the users system package manager slower, or install
unaudited packages (if they have a company mirror), but risk false
negatives, which should not harm too much, as the user should get
a relatively straight forward error message from `apt` then.

To be honest I am not really satisfied with this solution, but I have
not come up with a better one.

This is also the reason why I invoke `add-apt-repository` directly in
this case rather than using ansibles `apt_repository` in the hope that
this script will keep the user preferences.